requirements_celery.txt
celery==5.3.6
redis==5.0.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
PyMySQL==1.1.1
numpy==1.26.4
scipy==1.15.3
pandas==1.5.0
xarray==2023.12.0
matplotlib==3.10.3
Pillow==11.2.1
tqdm==4.64.1
Shapely==1.8.4
spatial==0.2.0
pycomlink==0.4.1
tensorflow==2.15.0
torch==2.5.1
requests==2.32.4
influxdb-client==1.48.0
urllib3==1.26.12
setuptools==75.6.0